Remove CarOwnerProfileService class
The `CarOwnerProfileService` class has been completely removed from the `Korik.Application` namespace. This includes the deletion of all associated using directives, the class definition, and its constructor. The class previously implemented the `ICarOwnerProfileService` interface and extended `GenericService<CarOwnerProfile>`. It also contained a private field `_repository` of type `IGenericRepository<CarOwnerProfile>`, which was initialized in the constructor.