Add 'ReadyForPickup' status and notification type
Introduce a new booking status, `ReadyForPickup`, to the
`BookingStatus` enumeration in `Booking.cs`. This status
is part of the booking lifecycle and triggers a notification
to users when their car is ready for pickup.

Update `UpdateBookingStatusRequest.cs` to send a
notification message, "Your Car is Ready For Pickup!",
with the notification type `NotificationType.CarReadyForPickup`
when the `ReadyForPickup` status is set.

Add a new notification type, `CarReadyForPickup`, to
`NotificationType.cs` to support the new booking status.